+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changes:
 - Support the money-open-exchange-rates gem.
 - Drop support for Rails < 6.1.
 - Database transaction when saving rates.
+- Add named argument for the list of currencies.
 
 Fix:
 - A nil bank defaults to EuCentralBank.